Article rename and final tweaks
