+++ client/demos/2020/avif-has-landed/compare/ZoomableTwoUp.tsx
