+++ client/demos/2020/avif-has-landed/compare/web-components.d.ts
