+++ shared/demos/2020/avif-has-landed/DecodedImg/utils.ts
