+++ static-build/index.tsx
@@ -79,9 +79,9 @@ toOutput['posts.rss'] = feed.atom1();
 // Demos
 import avifCompareDemoScript, {
   imports as avifCompareDemoScriptImports,
-} from 'client-bundle:client/demos/2020/avif-is-exciting/compare';
-import avifCompareDemoStyles from 'css-bundle:client/demos/2020/avif-is-exciting/compare/styles.css';
-toOutput['2020/avif-is-exciting/demos/compare/index.html'] = renderPage(
+} from 'client-bundle:client/demos/2020/avif-has-landed/compare';
+import avifCompareDemoStyles from 'css-bundle:client/demos/2020/avif-has-landed/compare/styles.css';
+toOutput['2020/avif-has-landed/demos/compare/index.html'] = renderPage(
   <ClientDemo
     title="Image comparison"
     script={avifCompareDemoScript}
@@ -92,9 +92,9 @@ toOutput['2020/avif-is-exciting/demos/compare/index.html'] = renderPage(
 
 import avifLoadingDemoScript, {
   imports as avifLoadingDemoScriptImports,
-} from 'client-bundle:client/demos/2020/avif-is-exciting/loading';
-import avifLoadingDemoStyles from 'css-bundle:client/demos/2020/avif-is-exciting/loading/styles.css';
-toOutput['2020/avif-is-exciting/demos/loading/index.html'] = renderPage(
+} from 'client-bundle:client/demos/2020/avif-has-landed/loading';
+import avifLoadingDemoStyles from 'css-bundle:client/demos/2020/avif-has-landed/loading/styles.css';
+toOutput['2020/avif-has-landed/demos/loading/index.html'] = renderPage(
   <ClientDemo
     title="Slow loading images"
     script={avifLoadingDemoScript}
