+++ static-build/posts/2020/09/avif-has-landed/demos/car-good.avif
